task2-KAN-42 add examples View
