--- src/components/Example2/Example2.tsx
@@ -1,5 +1,0 @@
-const Example2: React.FC = () => {
-  return <h2>Example2</h2>;
-};
-
-export default Example2;
