--- src/components/Example3/Example3.tsx
@@ -1,5 +1,0 @@
-const Example3: React.FC = () => {
-  return <h2>Example3</h2>;
-};
-
-export default Example3;
