+++ src/components/certificates/CertificateTable.tsx
@@ -1,34 +1,15 @@
-import { Certificate, CertificateType } from '../../types/types';
+import { Certificate } from '../../types/types';
 import './certificateTable.css';
 
-const certificates: Certificate[] = [
-  {
-    id: 1,
-    supplier: 'DAIMLER AG, 1, Berlin',
-    certificateType: CertificateType.PermissionOfPrinting,
-    validFrom: new Date('2017-08-21'),
-    validTo: new Date('2017-08-26'),
-  },
-  {
-    id: 2,
-    supplier: 'ANDEMIS GmbH, 1, Stuttgart',
-    certificateType: CertificateType.OHSAS18001,
-    validFrom: new Date('2017-08-18'),
-    validTo: new Date('2017-08-24'),
-  },
-  {
-    id: 3,
-    supplier: 'ANDEMIS GmbH, 1, Stuttgart',
-    certificateType: CertificateType.PermissionOfPrinting,
-    validFrom: new Date('2017-10-14'),
-    validTo: new Date('2017-10-20'),
-  },
-];
+type CertificateTableProps = {
+  certificates: Certificate[];
+};
 
-const CertificatesTable: React.FC = () => {
+const CertificatesTable: React.FC<CertificateTableProps> = ({
+  certificates,
+}) => {
   return (
     <section className="table">
-      <h2 className="table__title">Example1</h2>
       {certificates.length > 0 ? (
         <section className="table__container">
           <table className="table__content">
