+++ src/pages/Example2/Example2View.tsx
@@ -1,0 +1,5 @@
+const Example2View: React.FC = () => {
+  return <h2>Example2</h2>;
+};
+
+export default Example2View;
