+++ src/pages/Example3/Example3View.tsx
@@ -1,0 +1,5 @@
+const Example3View: React.FC = () => {
+  return <h2>Example3</h2>;
+};
+
+export default Example3View;
